feat: allow dynamic loop and throttle args (#44)
## Checklist

<!--
For each bullet, ensure your pr meets the criteria and write a note
explaining how this PR relates. Mark them as complete as they are done.
All top-level checkboxes should be checked regardless of their relevance
to the pr with a note explaining whether they are relevant or not.
-->

- [X] Code conforms to the [Elixir
Styleguide](https://github.com/christopheradams/elixir_style_guide)

## Problem

Sometimes we want dynamic throttle time depending on the args

## Details

This allows that.

---------

Co-authored-by: Jonathan Thomas <5272593+jondthomas@users.noreply.github.com>